old api same, new api excelParser.all()
old api of require('excelParser'); works the same, and now no longer
breaks when parsing excel files without xmlShareString.xml (ie excel
files without any string values).

new api of require('excelParser').all is a function that takes the same
inputs as before, but now returns an array of arrays of objects. each
object is a cell with key:values v:value and f:formula. all of the
properties are strings.

both api's are tested and working.

required extensive rewrite using adm-zip and bluebird instead of
node-promise, unzip and streams. the unzip library had a number of
issues and made it difficult to change course when no sharedString.xml
was piped through.